BETA: Add missing file update for [6650]

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/util/FloodProtectorConfig.java
@@ -47,7 +47,7 @@ public final class FloodProtectorConfig
 	/**
 	 * For how long should the char/account be punished.
 	 */
-	public int PUNISHMENT_TIME;
+	public long PUNISHMENT_TIME;
 	
 	/**
 	 * Creates new instance of FloodProtectorConfig.
